#3023 nonce added

+++ templates/features.php
@@ -7845,7 +7845,8 @@ function ampforwp_webp_featured_image() {
 }
 add_action('wp_ajax_ampforwp_notice_delete','ampforwp_notice_delete');
 function ampforwp_notice_delete(){
-	if ( current_user_can( 'install_plugins' ) || current_user_can( 'update_plugins' ) ) {
+	$wizard_nonce = $_REQUEST['wizard_nonce'];
+	if ( wp_verify_nonce( $wizard_nonce, 'wizard_nonce' ) && current_user_can( 'install_plugins' ) || current_user_can( 'update_plugins' ) ) {
 	  delete_transient( 'ampforwp_admin_notice_transient');
 	}
 }
